+++ apple-codesign/src/bundle_signing.rs
@@ -9,7 +9,7 @@ use {
         code_directory::CodeDirectoryBlob,
         code_requirement::RequirementType,
         code_resources::{CodeResourcesBuilder, CodeResourcesRule},
-        embedded_signature::{Blob, BlobData, CodeSigningSlot, DigestType},
+        embedded_signature::{Blob, BlobData, CodeSigningSlot},
         error::AppleCodesignError,
         macho::AppleSignable,
         macho_signing::MachOSigner,
@@ -495,14 +495,6 @@ impl SingleBundleSigner {
             let signer = MachOSigner::new(&macho_data)?;
 
             let mut settings = settings.clone();
-
-            // Framework bundles appear to use SHA-1 as the digest in the primary code directory
-            // and SHA-256 in the code directory at the 1st alternative code directory slot.
-            // Hardcoding the digest types here is likely subtly wrong...
-            if self.bundle.package_type() == BundlePackageType::Framework {
-                settings.set_digest_type(DigestType::Sha1);
-                settings.add_extra_digest(SettingsScope::Main, DigestType::Sha256);
-            }
 
             // The identifier for the main executable is defined in the bundle's Info.plist.
             if let Some(ident) = self